requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-database-url==1.0.3
django-filter==21.1
django-heroku==0.3.1
django-registration==3.1
django-registration-redux==2.10
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
